+++ src/features/gameState/GameState.ts
@@ -1,6 +1,7 @@
 interface GameState {
   inProgress: { value: boolean };
   isLevelLoading: { value: boolean };
+  hasWon?: { value: boolean };
 }
 
 export default GameState;