+++ changelog.md
@@ -6,6 +6,7 @@
     - Fix React warning in ``SingleSetting.tsx``
     - Configure project to read environment variables
     - Update dependencies
+    - Add ``hasWon`` property to Redux store
 
 # 0.6.2 Beta - 21/09/2024 17:00 UTC+1
 * Organised md documents into sections